move validation.js to middleware folder and update index.js accordingly
